requirements.txt
tqdm == 4.46.1
numpy == 1.18.5
argparse == 1.4.0
attrdict == 2.0.1
PyYAML == 5.3.1
matplotlib == 3.2.1
opencv-python == 4.2.0.34
scikit-learn == 0.23.1
tensorflow == 1.15.0
keras == 2.3.1
